refactor: 사용자 관심사 코드 리팩터링
